interview-prep/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
openai==1.57.0
python-dotenv==1.0.1
httpx==0.28.1
email-validator==2.2.0
asyncpg==0.30.0
redis==5.2.0
celery==5.4.0
python-dateutil==2.8.2
pytz==2024.1

